# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAPhoto
  description: |-
    Keep in touch with your friends by sharing photos of special moments, thanks to WASAPhoto! 
    You can upload your photo directly from your PC, and they will be visible to everyone following you
  version: 1.0.0

#============================================TAGS PART==============================================#
...
security:
  - bearerAuth: []
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      description: Bearer token for the WASAPhoto API.
      scheme: bearer

#============================================SCHEMAS PART==============================================#
  schemas:
...
    Code-Identifier:
      title: Code-random
      type: object
      description: |-
        General random generated string, which will be used as a unique identifier.
        It is a unique code that identifier uniquely:
          - the user
          - the photo
          - the comment
      properties:
        identifier: 
          type: string
          description: unique identifier generatedr randomly
          pattern: '^.*?$'
          minLength: 64
          maxLength: 64
          example: "115981bedf5cc6956330a1e4c394526a8839e8a3ec3adf0fre3eedd59c759a58"
      
    
#_____________________________USER FIELD____________________________________________
#-------------------------------------------------------------------------------- 
...
    UserID:
      title: UserID
      description: It's an object that contains an identifier code of user
      type: object
      properties:
        code-user: 
          $ref: "#/components/schemas/Code-Identifier"
      
#--------------------------------------------------------------------------------
...
    CommentID:
      title: Comment ID
      description: It's a object that contains an identifier code of comment
      type: object
      properties:
        code-comment:
          $ref: "#/components/schemas/Code-Identifier"
        
#--------------------------------------------------------------------------------  
    Comment:
      title: Comment
      description: |-
        Comment is an object which are strings writing by Users. It's composed by:
          - comment id
          - date and time that he/she posted the comment
          - the body string of the comment
          - owner comment
      type: object
      properties:
        comment-id:
          $ref: "#/components/schemas/CommentID"

        photo-id: 
          $ref: "#/components/schemas/ImageID"
        
        date-time:
          type: string
          description: Date and Time that was published the comment
          format: date-time
          pattern: ^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}Z$
          minLength: 20
          maxLength: 20
          readOnly: true
        
        text:
          type: string
          description: body string of the comment
          pattern: ^[^\\]{1,1024}$
          minLength: 1
          maxLength: 1024
        
        user-id: 
          $ref: "#/components/schemas/UserID"

        from-user: 
          $ref: "#/components/schemas/Username/properties/username-string"

        
      



#_____________________________LIKE FIELD____________________________________________
#--------------------------------------------------------------------------------
...
    ImageID:
      title: ImageID
      description: It's an object that contains an identifier code of image
      type: object
      properties:
        code-image: 
          $ref: "#/components/schemas/Code-Identifier"
          
#-------------------------------------------------------------------------------- 
...
    Error:
      title: Error
      type: object
      description: |-
        An error message that contains:
          - Code 
          - Message
      properties:
        code-error:
          type: integer
          description: The error code
          minimum: 3
          maximum: 3

        message-error:
          type: string
          description: The error message
          minLength: 1
          maxLength: 256
          pattern: ^[^\\]{1,256}$

#============================================PARAMETERS PART==============================================#
  parameters:
    user-id:
      name: user-id
      in: path
      description: The ID of a user to get information about him/her
      required: true
      schema: 
        $ref: "#/components/schemas/Username/properties/username-string"
      
#--------------------------------------------------------------------------------
...
    photo-id:
      name: photo-id
      in: path
      description: The ID photo identifier
      required: true
      schema: 
        $ref: "#/components/schemas/Code-Identifier/properties/identifier"
        
#--------------------------------------------------------------------------------
    comment-id:
      name: comment-id
      in: path
      description: The ID comment identifier
      required: true
      schema: 
        $ref: "#/components/schemas/Code-Identifier/properties/identifier"

#--------------------------------------------------------------------------------
...
paths:
...
  /users/{user-id}/homepage/{photo-id}/comments:
    parameters: 
      - $ref: "#/components/parameters/user-id"
      - $ref: "#/components/parameters/photo-id"
    # methods
    post:
      tags: ["comment", "photo"]
      summary: Post comment to the photo
      description: the user comments a photo so the user must be authenticated
      operationId: commentPhoto
      
      requestBody:
        description: the comment to post
        required: true
        content:
          application/json:
            schema: 
              $ref: "#/components/schemas/Comment"
            
      responses:
        "201":
          description: "The photo has been commented"
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Comment"

        "400":
          description: Invalid photo-id/user-id
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Error"

        "401":
          description: |-
            The requested user-id is invalid, no authentication information
            has been provided
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Error"

        "403":
          description: The comment is forbidden
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Error"

        "404":
          description: The resources are not found
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Error"

        "500":
          description: Internal Server Error
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Error"
  /users/{user-id}/homepage/{photo-id}/comments/{comment-id}:
  # Metodi: uguali ai like
    parameters: 
      - $ref: "#/components/parameters/user-id"
      - $ref: "#/components/parameters/photo-id"
      - $ref: "#/components/parameters/comment-id"
    # method

    delete:
      tags: ["comment", "photo"]
      summary: Delete comment the Photo
      description: the user uncomments a photo so the user must be authenticated
      operationId: uncommentPhoto
      
      responses:
        "204":
          description: "The photo has been uncomment"

        "400":
          description: Invalid photo-id/user-id/comment-id
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Error"
        "401":
          description: |-
            The requested user-id is invalid, no authentication information
            has been provided
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Error"

        "403":
          description: The uncomment is forbidden
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Error"

        "404":
          description: The comment-id is not found
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Error"

        "500":
          description: Internal Server Error
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Error"


#_____________________________PROFILE__________________________________
#----------------------------------------------------------------------
...